signTxMint: update for stax
