+++ src/signTxMint_ui.h
@@ -1,0 +1,37 @@
+#ifndef H_CARDANO_APP_SIGN_TX_MINT_UI
+#define H_CARDANO_APP_SIGN_TX_MINT_UI
+
+enum {
+	HANDLE_MINT_TOP_LEVEL_DATA_DISPLAY = 9200,
+	HANDLE_MINT_TOP_LEVEL_DATA_RESPOND,
+	HANDLE_MINT_TOP_LEVEL_DATA_INVALID,
+};
+
+void signTxMint_handleTopLevelData_ui_runStep();
+
+enum {
+	HANDLE_ASSET_GROUP_STEP_DISPLAY = 9300,
+	HANDLE_ASSET_GROUP_STEP_RESPOND,
+	HANDLE_ASSET_GROUP_STEP_INVALID,
+};
+
+void signTxMint_handleAssetGroup_ui_runStep();
+
+enum {
+	HANDLE_TOKEN_STEP_DISPLAY_NAME = 9400,
+	HANDLE_TOKEN_STEP_DISPLAY_AMOUNT,
+	HANDLE_TOKEN_STEP_RESPOND,
+	HANDLE_TOKEN_STEP_INVALID,
+};
+
+void signTxMint_handleToken_ui_runStep();
+
+enum {
+	HANDLE_CONFIRM_STEP_FINAL_CONFIRM = 9500,
+	HANDLE_CONFIRM_STEP_RESPOND,
+	HANDLE_CONFIRM_STEP_INVALID,
+};
+
+void signTxMint_handleConfirm_ui_runStep();
+
+#endif // H_CARDANO_APP_SIGN_TX_MINT_UI
